+++ src/components/integrationsComp/integrationEvents/integrationEvents.js
@@ -1,9 +1,14 @@
+import DoubleAppEvents from './doubleAppEvents/doubleAppEvents';
 import SingleAppsEvents from './singleAppEvents/singleAppsEvents';
 
-export default function IntegrationEvents({ pluginData }) {
+export default function IntegrationEvents({ plugins, pathArray }) {
     return (
         <>
-            <SingleAppsEvents pluginData={pluginData} />
+            {pathArray?.length < 2 ? (
+                <SingleAppsEvents pluginData={plugins} />
+            ) : (
+                <DoubleAppEvents pathArray={pathArray} plugins={plugins} />
+            )}
         </>
     );
 }
