+++ src/components/integrationsComp/integrationsComp.js
@@ -17,14 +17,16 @@ export default function IntegrationsComp({
     blogs,
     usecases,
     getStartedData,
+    pathArray,
+    hideApps,
 }) {
     return (
         <>
             <IntegrationsHero combinationData={combinationData} pluginData={pluginData} />
 
-            {type !== 'doubleApp' && <IntegrationsApps pluginData={pluginData} showCategories={false} />}
+            {type !== 'doubleApp' && !hideApps && <IntegrationsApps pluginData={pluginData} showCategories={false} />}
 
-            {pluginData && <IntegrationEvents pluginData={pluginData} />}
+            {pluginData && <IntegrationEvents plugins={pluginData} pathArray={pathArray} />}
 
             {usecases?.length > 0 && (
                 <div className="container mx-auto py-24">
@@ -39,9 +41,9 @@ export default function IntegrationsComp({
                 </div>
             )}
 
-            {faqData && faqName && <FAQSection faqData={faqData} faqName={`[singleApp]`} />}
+            {faqData && faqName && <FAQSection faqData={faqData} faqName={faqName} />}
 
-            <IntegrationsAbout pluginData={pluginData} />
+            <IntegrationsAbout plugins={pluginData} />
 
             <IntegrationFooter getStartedData={getStartedData} />
         </>
