+++ src/components/integrationsComp/integrationsHero/integrationsHero.js
@@ -11,57 +11,59 @@ export default function IntegrationsHero({ combinationData, pluginData }) {
     const [mode, setMode] = useState('dark');
 
     useEffect(() => {
-        if (pluginData?.brandcolor) {
-            setNewBrandColor(pluginData.brandcolor);
+        if (pluginData[0]?.brandcolor) {
+            setNewBrandColor(pluginData[0].brandcolor);
         }
-    }, [pluginData?.brandcolor]);
+    }, [pluginData[0]?.brandcolor]);
 
     useEffect(() => {
         setMode(GetColorMode(newBrandColor));
     }, [newBrandColor]);
 
-    if (!pluginData) {
-        return null;
-    }
-
     const integrationText =
-        pluginData.name === pluginData.name
-            ? `Create integrations between ${pluginData.name} and your favorite App`
-            : `Create integrations between ${pluginData.name} and ${pluginData.name}`;
+        pluginData.length == 1
+            ? `Create integrations between ${pluginData[0]?.name} and your favorite App`
+            : `Create integrations between ${pluginData[0]?.name} and ${pluginData[1]?.name}`;
 
     return (
         <>
             <div style={{ backgroundColor: `${newBrandColor}` }} className="py-12">
                 <div className={`container flex flex-col md:gap-12 gap-6`}>
                     <div
-                        className={`${styles.plugin_name} md:px-8 md:p-3 py-1 px-4 rounded-md flex items-center gap-8 w-fit shadow-black shadow-sm`}
+                        className={`${styles.plugin_name} md:px-8 md:p-3 py-1 px-4 rounded-md  flex items-center gap-8 w-fit`}
                     >
-                        <div className="flex items-center gap-3 w-fit">
-                            {pluginData.iconurl && (
-                                <Image
-                                    src={pluginData.iconurl}
-                                    width={40}
-                                    height={40}
-                                    className="h-auto w-[30px] md:w-[36px]"
-                                    alt={pluginData.name || 'plugin icon'}
-                                />
-                            )}
-                            <div className="flex flex-col">
-                                {pluginData.name && <p className="md:text-2xl text-lg font-bold">{pluginData.name}</p>}
-                                {pluginData.category && (
-                                    <p className="text-[14px] uppercase text-gray-400">
-                                        {pluginData.category.join(', ')}
-                                    </p>
-                                )}
-                            </div>
-                        </div>
+                        {pluginData.length &&
+                            pluginData.map((plug, index) => {
+                                return (
+                                    <div key={index} className="flex items-center gap-3  w-fit">
+                                        {index > 0 && <span className="text-3xl mx-4"> + </span>}
+                                        {plug?.iconurl && (
+                                            <Image
+                                                src={plug.iconurl}
+                                                width={40}
+                                                height={40}
+                                                className="h-auto w-[30px] md:w-[36px]"
+                                                alt="notion"
+                                            />
+                                        )}
+                                        <div className="flex flex-col">
+                                            {plug?.name && (
+                                                <p className="md:text-2xl text-lg font-bold">{plug?.name}</p>
+                                            )}
+                                            {plug?.category && (
+                                                <p className="text-[14px] uppercase text-gray-400">{plug?.category}</p>
+                                            )}
+                                        </div>
+                                    </div>
+                                );
+                            })}
                     </div>
                     <h2
                         className={`lg:text-6xl md:text-4xl text-3xl font-bold ${mode === 'dark' ? 'text-white' : 'text-accent'}`}
                     >
                         {integrationText}
                     </h2>
-                    {pluginData.events?.length ? (
+                    {pluginData[0]?.events?.length ? (
                         <>
                             {combinationData?.combinations?.length > 0 ? (
                                 <ComboGrid combos={combinationData} mode={mode} />
@@ -70,7 +72,7 @@ export default function IntegrationsHero({ combinationData, pluginData }) {
                             )}
                         </>
                     ) : (
-                        <NoDataGrid plugin={pluginData} mode={mode} />
+                        <NoDataGrid plugin={pluginData[0]} mode={mode} />
                     )}
                 </div>
             </div>
